requirements.txt
streamlit
openai
python-dotenv
pdfplumber
python-docx
faiss-cpu
sentence-transformers
fpdf